feat: migrate Pages Router to App Router (#168)
- Add 'use client' to Nav.tsx (uses useState/useEffect)
- Create client components: FontAwesomeSetup, GoogleAnalytics, ZennEmbedLoader
- Create app/layout.tsx combining _app.tsx + _document.tsx with Next.js Metadata API
- Create app/page.tsx, app/blogs/[id]/page.tsx, app/pages/blogs/[id]/page.tsx, app/pages/tags/[id]/page.tsx as async Server Components with generateStaticParams/generateMetadata
- Delete all pages/ router files and src/libs/next-seo.config.ts
- Uninstall next-seo; replace with built-in Metadata API
- Remove @/pages/* path alias from tsconfig.json and jest.config.js
