+++ app/page.tsx
@@ -1,0 +1,30 @@
+import React from 'react'
+import { Layout, Card, Pagination } from '@/src/index'
+import { tagsGetAllContents, blogsGetHeader, blogsGetTotalCount } from '@/src/libs/getContents'
+import type { blog, tag } from '@/src/index'
+
+export default async function Home() {
+  const [blogs, tags, blogsCount]: [blog[], tag[], number] = await Promise.all([
+    blogsGetHeader(),
+    tagsGetAllContents(),
+    blogsGetTotalCount(),
+  ])
+
+  return (
+    <Layout latestBlogs={blogs} tags={tags}>
+      <div className="mb-2 flex items-center justify-between">
+        <h2 className="font-head text-xl text-base-content md:text-2xl">新着記事</h2>
+      </div>
+
+      {blogs.map((blog) => (
+        <React.Fragment key={blog.id}>
+          <div className="mb-2">
+            <Card data={blog} />
+          </div>
+        </React.Fragment>
+      ))}
+
+      <Pagination totalCount={blogsCount} pageNumber={1} />
+    </Layout>
+  )
+}
