+++ package.json
@@ -40,7 +40,6 @@
     "gray-matter": "^4.0.3",
     "jsdom": "27.4.0",
     "next": "16.1.6",
-    "next-seo": "7.1.0",
     "react": "19.2.4",
     "react-dom": "19.2.4",
     "sass": "1.97.3",
