--- pages/_app.tsx
@@ -1,67 +1,0 @@
-import '../styles/globals.css'
-import { config, library } from '@fortawesome/fontawesome-svg-core'
-import {
-  faHome,
-  faEdit,
-  faUser,
-  faEnvelope,
-  faWindowMaximize,
-  faArrowRight,
-  faCalendarPlus,
-} from '@fortawesome/free-solid-svg-icons'
-import {
-  faYoutube,
-  faTwitterSquare,
-  faInstagramSquare,
-  faFacebookSquare,
-  faTiktok,
-} from '@fortawesome/free-brands-svg-icons'
-import '@fortawesome/fontawesome-svg-core/styles.css'
-
-config.autoAddCss = false
-library.add(
-  faHome,
-  faEdit,
-  faUser,
-  faEnvelope,
-  faWindowMaximize,
-  faArrowRight,
-  faYoutube,
-  faTwitterSquare,
-  faInstagramSquare,
-  faFacebookSquare,
-  faTiktok,
-  faCalendarPlus,
-)
-
-// gtag
-import { useEffect } from 'react'
-import { AppProps } from 'next/app'
-import { useRouter } from 'next/router'
-import { generateDefaultSeo } from 'next-seo/pages'
-import { gtag, SEO } from '@/src/index'
-
-export default function MyApp({ Component, pageProps }: AppProps) {
-  const router = useRouter()
-
-  useEffect(() => {
-    const handleRouteChange = (url: string) => {
-      gtag.pageview(url)
-    }
-    router.events.on('routeChangeComplete', handleRouteChange)
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange)
-    }
-  }, [router.events])
-
-  useEffect(() => {
-    import('zenn-embed-elements')
-  }, [])
-
-  return (
-    <>
-      {generateDefaultSeo(SEO)}
-      <Component {...pageProps} />
-    </>
-  )
-}
