--- pages/blogs/[id].tsx
@@ -1,130 +1,0 @@
-import React from 'react'
-import { GetStaticProps, GetStaticPaths } from 'next'
-import Link from 'next/link'
-import Image from 'next/image'
-import { generateNextSeo } from 'next-seo/pages'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { JSDOM } from 'jsdom'
-import { Layout, Date, SEO } from '@/src/index'
-import { blogsGetAllHeader, blogGetContent } from '@/src/libs/getContents'
-import type { blogData, tableOfContents } from '@/src/index'
-import Style from '@/styles/blog.module.scss'
-import base64url from 'base64url'
-import markdownToHtml from 'zenn-markdown-html'
-import 'zenn-content-css'
-
-export const getStaticPaths: GetStaticPaths = async () => {
-  const blogsData = await blogsGetAllHeader()
-  const paths = blogsData.map((blogData) => `/blogs/${blogData.id}`)
-  return { paths, fallback: 'blocking' }
-}
-
-export const getStaticProps: GetStaticProps = async (context) => {
-  const id = context.params?.id
-  const blog = id !== undefined && !Array.isArray(id) ? await blogGetContent(id) : await blogGetContent('')
-
-  // markdownからhtmlに変換
-  const html = await markdownToHtml(blog.body, {
-    embedOrigin: 'https://embed.zenn.studio',
-  })
-
-  // htmlでパースできるようにする
-  const dom = new JSDOM(html)
-
-  // 目次
-  const headings = Array.from(dom.window.document.querySelectorAll<HTMLElement>('h1, h2, h3, h4, h5, h6'))
-  const tableOfContents = headings.map((element) => ({
-    text: element.textContent,
-    id: element.getAttribute('id'),
-    name: element.nodeName,
-  }))
-
-  return {
-    props: {
-      blog: blog,
-      highlightedBody: dom.window.document.body.outerHTML,
-      tableOfContents: tableOfContents,
-    },
-  }
-}
-
-const baseURL: string = process.env.NEXT_PUBLIC_BASE_URL ?? ''
-
-type props = {
-  blog: blogData
-  highlightedBody: string
-  tableOfContents: tableOfContents[]
-}
-
-export default function Blog({ blog, highlightedBody, tableOfContents }: props) {
-  const width = 1200
-  const ogpBaseImage = `${baseURL}/ogp/article_1200x630.png`
-  const ogpTitle = `https://assets.imgix.net/~text?txtsize=48&txt-color=1F2937&w=${
-    width - 80
-  }&txt-align=middle&txtfont=Hiragino%20Sans%20W6&txt-track=2`
-  return (
-    <>
-      {generateNextSeo({
-        ...SEO,
-        title: blog.title,
-        titleTemplate: '%s - Offiter',
-        description: blog.description,
-        openGraph: {
-          type: 'article',
-          url: `${baseURL}/blogs/${blog.id}`,
-          title: blog.title,
-          description: blog.description,
-          images: [
-            {
-              url: `${ogpBaseImage}?blend64=${base64url(
-                `${ogpTitle}&txt64=${base64url(blog.title)}`,
-              )}&blend-mode=normal&blend-align=top,left&blend-x=40&blend-y=100`,
-              height: 630,
-              width: 1200,
-              alt: 'Og Image Alt',
-            },
-          ],
-        },
-      })}
-      <Layout blogDetails={blog} tableOfContents={tableOfContents}>
-        <div className="lg:card lg:bg-base-100 lg:p-2 lg:shadow-md">
-          <div className="mt-4 flex items-center justify-between lg:hidden">
-            <div className="flex flex-col">
-              <span className="text-xs font-light text-base-content/70">
-                <FontAwesomeIcon icon="calendar-plus" fixedWidth />
-                {Date(blog.createdAt)}
-              </span>
-              <span className="text-xs font-light text-base-content/70">
-                <FontAwesomeIcon icon="edit" fixedWidth />
-                {Date(blog.updatedAt)}
-              </span>
-            </div>
-
-            <div className="flex items-center justify-end">
-              {blog.tags.map((tag) => (
-                <React.Fragment key={tag.id}>
-                  <Link
-                    href="/[tag]"
-                    as={`/${tag.id}`}
-                    className="badge badge-neutral ml-2 text-xs font-bold md:text-base"
-                  >
-                    <span>{tag.name}</span>
-                  </Link>
-                </React.Fragment>
-              ))}
-            </div>
-          </div>
-
-          <div className="mt-4 flex items-center justify-center">
-            <Image alt="" src={blog.image.url} width={blog.image.width ?? 1200} height={blog.image.height ?? 630} />
-          </div>
-
-          <div
-            className={`${Style.blog} znc mt-4 md:text-lg`}
-            dangerouslySetInnerHTML={{ __html: `${highlightedBody}` }}
-          />
-        </div>
-      </Layout>
-    </>
-  )
-}
