+++ src/components/Nav.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useState, useEffect } from 'react'
 import Link from 'next/link'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
