+++ src/components/ZennEmbedLoader.tsx
@@ -1,0 +1,10 @@
+'use client'
+import { useEffect } from 'react'
+
+export default function ZennEmbedLoader() {
+  useEffect(() => {
+    import('zenn-embed-elements')
+  }, [])
+
+  return null
+}
